+++ server/src/main/java/net/feedbacky/app/StartupValidator.java
@@ -46,7 +46,7 @@ public class StartupValidator {
     MAIL_SMTP_HOST("0.1.0-beta"), MAIL_SMTP_PORT("0.1.0-beta"), IMAGE_COMPRESSION_ENABLED("0.1.0-beta"), IMAGE_COMPRESSION_TYPE("0.1.0-beta"),
     IMAGE_COMPRESSION_CHEETAHO_API_KEY("0.1.0-beta"),
     MAIL_SENDGRID_API_KEY("0.2.0-beta"), MAIL_SENDGRID_API_BASE_URL("0.2.0-beta"),
-    SETTINGS_ALLOW_COMMENTING_CLOSED_IDEAS("0.5.0-beta");
+    SETTINGS_ALLOW_COMMENTING_CLOSED_IDEAS("0.5.0-beta"), SETTINGS_MAIL_LOGIN_ENABLED("1.0.0.beta.4");
 
     private final String sinceVersion;