+++ server/src/main/java/net/feedbacky/app/data/AboutFeedbackyData.java
@@ -20,5 +20,6 @@ public class AboutFeedbackyData {
   private List<LoginProvider.ProviderData> loginProviders;
   private List<FetchUserDto> serviceAdmins;
   private boolean closedIdeasCommenting;
+  private boolean mailLoginEnabled;
 
 }